refactor(desktop): share session deletion logic
Extract the duplicated delete-with-undo flow from the calendar session chip and top meeting timeline into a useDeleteSession hook.
